+++ src/bin/rage-keygen/main.rs
@@ -17,13 +17,14 @@ fn main() {
 
     let opts = AgeOptions::parse_args_default_or_exit();
 
-    let mut output = match file_io::OutputWriter::new(opts.output, file_io::OutputFormat::Text) {
-        Ok(output) => output,
-        Err(e) => {
-            error!("Failed to open output: {}", e);
-            return;
-        }
-    };
+    let mut output =
+        match file_io::OutputWriter::new(opts.output, file_io::OutputFormat::Text, 0o600) {
+            Ok(output) => output,
+            Err(e) => {
+                error!("Failed to open output: {}", e);
+                return;
+            }
+        };
 
     let sk = age::keys::SecretKey::generate();